+++ test/__init__.py
@@ -1,0 +1,2 @@
+URL = 'http://localhost:8080'
+#~ URL = 'https://tallersharedserver.herokuapp.com'
